+++ extensions/browser-history/src/components/actions.tsx
@@ -1,4 +1,4 @@
-import { Action, ActionPanel, Detail, Icon, openCommandPreferences } from "@raycast/api";
+import { Action, ActionPanel, Icon, openCommandPreferences } from "@raycast/api";
 import { openNewFirefoxTab, openNewTab } from "../actions";
 import { HistoryEntry, SupportedBrowsers } from "../interfaces";
 
@@ -35,6 +35,11 @@ export const HistoryItemAction = ({ entry: { title, url } }: { entry: HistoryEnt
         icon={"vivaldi-logo.png"}
         onAction={() => openNewTab(SupportedBrowsers.Vivaldi, url)}
       />
+      <ActionPanel.Item
+        title={"Open in Arc"}
+        icon={"arc-logo.svg"}
+        onAction={() => openNewTab(SupportedBrowsers.Arc, url)}
+      />
     </ActionPanel.Section>
   </ActionPanel>
 );
@@ -51,28 +56,3 @@ export function ActionOpenPreferences() {
     </ActionPanel.Section>
   );
 }
-
-const permissionErrorMarkdown = `## Raycast needs full disk access in order to display your Safari bookmarks.
-
-![Full Disk Access Preferences Pane](https://i.imgur.com/3SAUwrx.png)
-
-1. Open the **Security & Privacy** Preferences pane and select the **Privacy** tab
-2. Select **Full Disk Access** from the list of services
-3. Click the lock icon in the bottom left corner to unlock the interface
-4. Enter your macOS administrator password
-5. Drag and Drop the icon for the **Raycast** application into the list as seen above`;
-
-export const PermissionErrorDetail = () => (
-  <Detail
-    markdown={permissionErrorMarkdown}
-    actions={
-      <ActionPanel>
-        <Action.Open
-          title="Open System Preferences"
-          icon={Icon.Gear}
-          target="x-apple.systempreferences:com.apple.preference.security?Privacy_AllFiles"
-        />
-      </ActionPanel>
-    }
-  />
-);