+++ extensions/browser-history/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Browser History Changelog
 
+## [Add Arc Browser] - 2022-12-16
+
+- Added Arc history search support.
+
 ## [Initial Version] - 2022-12-13
 
 - Added Chrome, Firefox, Safari, Edge, Brave and Vivaldi history search support.